pyproject.toml
[tool.poetry]
name = "wwf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
pytesseract = "^0.3.10"
numpy = "^1.23.3"
colorthief = "^0.2.1"
opencv-python = "^4.6.0.66"
imutils = "^0.5.4"
requests = "^2.28.1"
pytest-playwright = "^0.3.0"
asyncio = "^3.4.3"
PyAutoGUI = "^0.9.53"
PyGetWindow = "^0.0.9"
PyWinCtl = "^0.0.38"
json5 = "^0.9.10"
pynput = "^1.7.6"
psutil = "^5.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
